% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elif.R
\name{elif}
\alias{elif}
\title{elif
Generate vectorized if-else logic with pleasing syntax}
\usage{
elif(..., env = parent.frame())
}
\arguments{
\item{...}{Expressions separated by commas mapping predicates to values. The
\code{->} operator reminiscent to pattern matching in other languages. The last
expression must be a default value or an error will be thrown.}

\item{env}{Where the resultant expression should be evaluated. Defaults
to the \code{parent.frame()}.}
}
\value{
the result of the evaulated \code{ifelse} expression evaulated in \code{env}.
}
\description{
elif
Generate vectorized if-else logic with pleasing syntax
}
\examples{
\dontrun{
res <- with(mtcars, elif(
 cyl == 4 -> "a",
 cyl == 6 -> "b",
 cyl == 8 -> "c", "d"))
table(res, mtcars$cyl)
}
}
